refactor: Generator 객체를 패키지를 만들어 따로 관리
