+++ src/main/java/controller/BlackJackController.java
@@ -4,17 +4,17 @@ import static java.util.stream.Collectors.toList;
 
 import domain.Account;
 import domain.CardDeck;
-import domain.CardDeckGenerator;
 import domain.Dealer;
 import domain.DrawCommand;
 import domain.Name;
-import domain.ParticipantGenerator;
 import domain.Player;
 import domain.Players;
 import domain.Status;
 import dto.BlackJackResult;
 import dto.DrawnCardsInfo;
 import dto.ParticipantResult;
+import generator.CardDeckGenerator;
+import generator.ParticipantGenerator;
 import java.util.List;
 import service.BlackJackService;
 import view.InputView;
