+++ src/test/java/service/BlackJackServiceTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import domain.Account;
 import domain.Card;
 import domain.CardDeck;
-import domain.CardDeckGenerator;
 import domain.Dealer;
 import domain.DrawCommand;
 import domain.DrawnCards;
@@ -18,6 +17,7 @@ import domain.Type;
 import domain.Value;
 import dto.DrawnCardsInfo;
 import dto.ParticipantResult;
+import generator.CardDeckGenerator;
 import java.util.ArrayList;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
